Move Redmine client into RMT namespace
